refactor: organize components into logical folder structure
## Component Organization Changes
- Move navigation components to `layout/` folder
- Move UI/layout components to `ui/` folder
- Move SEO components to `common/` folder
- Move Instagram components to `social/` folder
- Create new `optimization/` folder for business logic components
- Update all import paths to use absolute imports from new locations
- Create/update index.ts files for proper exports

## Moved Components
**Layout/Navigation:**
- BreadcrumbNavigation.tsx → layout/
- CategoryNavigation.tsx → layout/
- CrossServiceNavigation.tsx → layout/
- MobileNavigation.tsx → layout/
- ServiceNavigationControls.tsx → layout/

**UI Components:**
- ContextAwareHeader.tsx → ui/
- ServicePageLayout.tsx → ui/
- LazyImage.tsx → ui/
- ThemeToggle.tsx → ui/

**SEO/Common:**
- ServicePageSEO.tsx → common/

**Social Media:**
- InstagramEmbed.tsx → social/
- InstagramOfficialEmbed.tsx → social/

**Business Logic:**
- JourneyOptimization.tsx → optimization/

## Cleanup
- Remove 5 temporary PerformanceGallery.tsx files (.tmp, .tmp2-5)
- Fix import paths in affected files
- Update ThemeToggle to use absolute imports
- Fix LazyImage import in PerformanceGallery

## Quality Assurance
- ✅ All imports updated and verified
- ✅ TypeScript compilation successful
- ✅ Build process successful
- ✅ No breaking changes to functionality
