+++ src/components/common/index.ts
@@ -8,6 +8,7 @@ export {
   useErrorHandler,
 } from './ErrorBoundary'
 export { default as SEOHead } from './SEOHead'
+export { default as ServicePageSEO } from './ServicePageSEO'
 export { usePageSEO } from '@/hooks/usePageSEO'
 export { generateMetaDescription, validateStructuredData } from '@/utils/seo'
 export {
