+++ src/components/layout/CategoryNavigation.tsx
