+++ src/components/layout/index.ts
@@ -1,0 +1,7 @@
+// Layout and Navigation components
+export { default as Navigation } from './Navigation'
+export { default as BreadcrumbNavigation } from './BreadcrumbNavigation'
+export { default as CategoryNavigation } from './CategoryNavigation'
+export { default as CrossServiceNavigation } from './CrossServiceNavigation'
+export { default as MobileNavigation } from './MobileNavigation'
+export { default as ServiceNavigationControls } from './ServiceNavigationControls'
