+++ src/components/optimization/index.ts
@@ -1,0 +1,2 @@
+// Optimization and Journey Components
+export { default as JourneyOptimization } from './JourneyOptimization'
