+++ src/components/social/index.ts
@@ -1,0 +1,4 @@
+// Social Media Integration Components
+export { default as SocialProof } from './SocialProof'
+export { default as InstagramEmbed } from './InstagramEmbed'
+export { default as InstagramOfficialEmbed } from './InstagramOfficialEmbed'
