+++ src/components/ui/ContextAwareHeader.tsx
